Rename io.ably to com.ably.
