+++ subscribing-example-app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="io.ably.tracking.example.subscriber">
+  package="com.ably.tracking.example.subscriber">
 
   <application
     android:allowBackup="true"
