+++ subscribing-example-app/src/main/java/com/ably/tracking/example/subscriber/MainActivity.kt
@@ -1,0 +1,11 @@
+package com.ably.tracking.example.subscriber
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+
+class MainActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+    }
+}
