--- subscribing-example-app/src/main/java/io/ably/tracking/example/subscriber/MainActivity.kt
@@ -1,11 +1,0 @@
-package io.ably.tracking.example.subscriber
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-    }
-}
