--- subscribing-example-java-app/src/androidTest/java/io/ably/tracking/example/javasubscriber/ExampleInstrumentedTest.java
@@ -1,26 +1,0 @@
-package io.ably.tracking.example.javasubscriber;
-
-import android.content.Context;
-
-import androidx.test.platform.app.InstrumentationRegistry;
-import androidx.test.ext.junit.runners.AndroidJUnit4;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-import static org.junit.Assert.*;
-
-/**
- * Instrumented test, which will execute on an Android device.
- *
- * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
- */
-@RunWith(AndroidJUnit4.class)
-public class ExampleInstrumentedTest {
-    @Test
-    public void useAppContext() {
-        // Context of the app under test.
-        Context appContext = InstrumentationRegistry.getInstrumentation().getTargetContext();
-        assertEquals("io.ably.tracking.example.javasubscriber", appContext.getPackageName());
-    }
-}
