+++ subscribing-example-java-app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="io.ably.tracking.example.javasubscriber">
+    package="com.ably.tracking.example.javasubscriber">
 
     <application
         android:allowBackup="true"
