--- subscribing-example-java-app/src/main/java/io/ably/tracking/example/javasubscriber/MainActivity.java
@@ -1,14 +1,0 @@
-package io.ably.tracking.example.javasubscriber;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-public class MainActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-    }
-}
